Move getUser functionality to context
