+++ src/context/github/GithubState.js
@@ -39,7 +39,21 @@ const GithubState = props => {
 		});
 	};
 
-	// Get user
+	// Get single user
+	const getUser = async login => {
+		setLoading();
+
+		const res = await axios.get(
+			`https://api.github.com/users/${login}?client_id=${
+				process.env.REACT_APP_GITHUB_CLIENT_ID
+			}&client_secret=${process.env.REACT_APP_GITHUB_CLIENT_SECRET}`
+		);
+
+		dispatch({
+			type: GET_USER,
+			payload: res.data
+		});
+	};
 
 	// Get repos
 
@@ -56,7 +70,8 @@ const GithubState = props => {
 				repos: state.repos,
 				loading: state.loading,
 				searchUsers,
-				clearUsers
+				clearUsers,
+				getUser
 			}}
 		>
 			{props.children}
